sys-io: refactor NET.
This is a simplification/refactoring of the NET subsystem.
Basically, initially it was overengineered to accommodate
any net stack implementation, but now it is basically tied
to smoltcp. IT makes it much more simpler and easiler to work
on/optimize.

The loopback devices is not available yet, will be added soonish.