vdso: refactor EventSource runtime.
I think this way is better a bit. Based on an LLM suggestion
(prompted), but substantially reworked (the LLM made a couple of errors,
renamed a couple of variables needlessly, and named a couple of
new variables in a way that I didn't like). Still feels a bit off
somehow...