Use `&mut [MaybeUninit<u8>]` for write-only buffers
Currently, the operational semantics are vague on whether the values
pointed to by a reference must be fully initialized. Tentatively, Rust
does *not* have full recursive validity for references
(https://github.com/rust-lang/unsafe-code-guidelines/issues/412), so
having a `&mut [u8]` slice would not guarantee that its elements are
initialized, but this is not official, so we cannot rely on it. More
strictly, the safety documentation for `core::slice::from_raw_parts_mut`
requires that it points to len number of initialized values. To be safe,
use `&mut [MaybeUninit<u8>]`.

This will enable passing possibly-uninitialized memory to these
`read`-like system calls, to implement `Read::read_buf` in the standard
library.